pyproject.toml
[project]
name = "bluesky-summarizer"
version = "0.1.0"
description = ""
authors = [{ name = "Stasiu and Marek" }]
requires-python = "~=3.11.0"
readme = "README.md"
dependencies = [
    "flask>=3.1.0,<4",
    "flask-cors>=6.0.0,<7",
    "atproto>=0.0.61,<0.0.62 ; python_version >= '3.11' and python_version < '3.14'",
    "google-genai>=1.14.0,<2",
    "retry>=0.9.2,<0.10",
    "python-dotenv>=1.0.0,<2",
    "libsql>=0.1.8,<0.2",
    "mailjet-rest>=1.5.1",
    "openai>=2.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
